requirements.txt
playwright
pyotp
pyzbar
Pillow
cryptography
tabulate
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
